Upgraded sprite transformations to SSE3 and disabled this code path.
It seems the code spends more time in store/load than is beneficial for all the
parallel computations to actually make a difference. The scalar implementation
is negligibly faster in this area, or perhaps we are simply bound by the
memory lanes on my laptop. Should investigate further when there is more time.