Android: Implemented OpenSL ES backend.
Both Sound and Stream should function now. The problem was Android's
implementation of the OpenSL ES specification, which is not only
outdated (1.0.1 <-> 1.1), but also has its own extensions. One cannot
choose to not use them. The implementation is Android-specific for
now. But then again, Android is the only platform that employs OpenSL.

Audio is not automatically cut when the app is in background like on
iOS, so this is still an issue to address.